refactor: migrate to newCustomFetch
